Update to bsnes v076r02 release.
Changelog:
- fixed a crashing bug when you toggle a cheat code and then load a
save state
- added paths.cfg
- bsnes now remembers the most recent path _per_ file-type, like
bsnes/Qt used to do
- re-added -s to Linux linker flags
- fixed crash with libsnes/SGB XML parsing

Okay, so the way paths work is pretty cool, but still a work in
progress. Every load type (SFC, BS, ST, GB, filter, shader) has its
own path string. This string keeps track of the last path you've
loaded a file from. The path is prefixed with a ":" to indicate that
it is not a manual path, but rather a recent path, and that subsequent
loads should update it with a new recent path. Specify a path without
the ":" and it will always load from that path, and never update the
path from subsequent path loads.

The support is currently only there for initial load paths and
remembering the BSX/ST/SGB BIOS file paths. A lot more work will be
needed for the rest.

Also, this is complicating my "don't reload the window if the path and
type is the same" code, which is what's needed to remember the last
selected file in the list. Man that's some annoying code to support :P