Update to bsnes v077r01 release.
Changelog:
- fixed auto joypad polling bug (fixes Super Star Wars start button
input)
- added pseudo-random class; so far it is RAM only [WRAM, APURAM,
VRAM, OAM, CGRAM]
- added new system configuration options to bsnes.cfg

The pseudo-random class is optional. For right now, I am defaulting it
to enabled. bsnes.cfg::snes.random = true.
You can of course turn it off if you want. This **will** break Death
Brade and Unnamed Euro Racer, no questions about it.
So I don't know if I want to leave this on or off by default. Leaving
it on will thwart emulator detection code and help to keep code that
relies on uninitialized memory from working, but leaving it off will
increase compatibility.

Interesting, huh? More accuracy typically increases compatibility, but
it has the potential at a certain point to decrease it as well.
Death Brade **will**, _beyond a shadow of a doubt_, break on certain
SNES units in the wild. Matching its pattern better will of course
help, but it's the same situation as Super Bonk.

So yeah, we'll see how things go before deciding on what to do with
v078 official.

Anyway, the randomness is a basic LFSR based on CRC32 (courtesy of
blargg), seed is 'byuu', set on power-on. I may also allow time(0) as
a seed. The seed is saved in the save state information. So once you
have a state, the randomness is predictable and repeatable.