Update to bsnes v077r05 release.
Please test this one as much as you guys can. Especially with toggling
NTSC/PAL TV mode, and using various filters, and running all games
that are known to be difficult to emulate.
The randomization could very well cause issues in some games if I
randomize anything that should be constant. I took an overly
aggressive approach.
We can verify it's a random issue by setting bsnes.cfg's snes.random
to false and see if that fixes the issue.
I want to leave snes.random enabled in v078, and I'd like to avoid any
major breakage if possible.

Changelog:
* fixed .nec and both .rtc file extensions (thanks krom)
* randomized most S-PPU registers, should trip up some broken homebrew
that does not initialize all registers
* randomization is now seeded with time(0) rather than 'byuu'
* SNES::interface.video_refresh() now always receives
{256,512}x{240,480}
* PPU/accuracy scanline 0 does not render the screen back color
anymore, fixes strange coloring look on first scanline in PAL TV mode
in non-overscan games
* disabled hires blending in PPU/compatibility; all three cores act
the same now

For seeding with time(0), this is basically to prevent any game from
relying on **any** specific value. It will catch more uninitialized
register access than a static seed. It has an added benefit of
stopping people from 'fingerprinting' the RAM data to detect bsnes,
although that's probably still possible because bsnes will have a
lower SNR for WRAM. We need to study many SNES WRAM dumps and come up
with a nicer random generator for it. At any rate, it will be
substantially harder to detect now.

For the video refresh, this is basically to simplify filters. It's a
pain in the ass dealing with all kinds of different sizes and odd
resolutions, and it just serves to complicate everything. Right now
I'm not actually clearing lines 225-240 when overscan is off, but that
should not be too difficult, and it makes a future TV-lite simulation
filter much easier to write.

There is one issue though, you can't detect overscan state anymore.
Hires is true if width == 512, and interlace is true is height == 480,
but overscan isn't exposed.
Hires is line-based, and always true on the accuracy renderer anyway.
Interlace and overscan are 'per-frame' things, you can toggle them but
they don't take effect until the next frame anyway.
So I think I need to extend video_refresh to expose overscan,
interlace, and field values. But I am hesitant to break BC on libsnes.
Maybe it's best to have snes_get_video_state() that returns a mask of
those settings or something.

What I really need to do is write up documentation on the internal
video architecture of bsnes.