Update to bsnes v078r02 release.
New S-SMP core is feature-complete, but still buggy.
It's good enough for perfect audio in Zelda 3 and Super Mario World,
but there are plenty of issues.
No audio in Bahamut Lagoon, deadlock in Earthworm Jim 2, etc.

With this core, bsnes/Performance runs about 3-5% faster than with the
old one. That won't seem like much, because the S-SMP is the least
demanding portion of the SNES.
blargg's SMP core netted me a 5-8% speedup the last time I tried it,
so I'm sure there's still room to speed things up.

The core is opcode-based, but has dummy op_io() calls (they compile to
nothing), so it is trivial to make it cycle-based if desired.
I'm not convinced that is necessary, but we shall see once we get the
opcode bugs ironed out.

If anyone wants to help, that'd be great. Just compare the
snes/smp/core opcode semantics to the snes/alt/smp/core semantics and
look for typos.
I have already ran synthetic tests to validate all cycle counts for
every opcode (sans conditional branches not taken), and to ensure that
all 256 opcodes are defined.
Probably best/fastest to rig up both SMP cores and perform some sort
of random execution looking for any differences.