Update to bsnes v078r07 release.
Would appreciate testing on any games with mosaic, especially Mode7
mosaic.
I have tested Super Turrican 2, Sim Earth, Contra III and SNES Test
Program.

This only applies to BG modes 0-6, and technically should not affect
Mode7 at all. I am not sure if Mode7 needs the same change made or
not, but given the way it fetches that could prove quite challenging.
I also simplified the background renderer a good bit. See eg the pixel
copy stuff in Background::run().

I've only fixed this in the accuracy renderer. I'm sorry, but the
compatibility renderer is a fucking mess. I haven't really touched it
in four or five years now.

Will probably just revert to the accuracy/SMP in the performance
profile for the next release since it's not being used otherwise.
People can toggle it on if they want to try it out.