Update to bsnes v079r02 release.
Added "unsigned link_run();" which acts as its own thread synchronized against the S-CPU. Specify the frequency in the configuration file.
I intend to prototype the Cx4 LLE openly using the link module, and that required timing support, so there we go.
It's very basic, and it synchronizes the CPU to the coprocessors and vice versa after every call to link_run().
Meaning performance won't be super exceptional at full 21MHz or higher, but then this is for prototyping only.
I didn't want to expose cothreading, yielding, calls back into bsnes' core, calls to sync up the S-CPU, etc.