Update to bsnes v080r01 release.
There was one unfortunate aspect of the S-DD1 module: you had to give
it the DMA length and a target buffer, and it would do the entire
decompression at once.
Real hardware would work by streaming the data byte by byte. So with
that, I went ahead and rewrote the code to handle byte-based
streaming.

This WIP is an important milestone for me personally. Up until now,
bsnes has always had code that was directly copy-pasted from other
authors.
With all of the DSP and Cx4 chips rewritten in LLE, and the SPC7110
algorithm already ported over from C, and archive decompression code
removed for a long time, the S-DD1 was the only module left like this.
It's obviously not that big of a deal. The code is basically still a
copy of the original. S-DD1 decomp from Andreas Naive, SPC7110 decomp
from neviksti, and S-DSP from blargg.
And the rest of the emulator is of course only possible because of
code and research before it, although everything else has no
resemblance at all to code before it.
The main advantage, really, is absolute code consistency. I always use
the same variant of K&R, for instance.
I dunno, I guess I just never really liked the "Build-a-Bear Workshop"
style of emulators, like is so prominent in the Genesis scene: "My new
Genesis emu (uses Starscream/Musashi 68K core, Marat Fayzullin's Z80
core, YM2612 core from Game_Music_Emu, VDP core from Gens, SVP core
from picodrive)", sorry, but you wrote a front-end, not an emulator :/

I also updated the SPC7110 decompression module: I merged the class
inside the SPC7110 class (not sure why it was separate before), and
replaced the morton lookup tables with for-loops.
The morton tables were added to be a tiny bit faster when I was more
interested in speed than code clarity.
It may be a tiny bit slower (or faster due to less L2 cache usage),
but you won't even notice an FPS drop, and it cuts out a good chunk of
code and some tables.
Lastly, I added pinput_poll() to video_refresh().
Forgot to remove Interface::input_poll() from the C++ side, will have
to do that later.