Update to bsnes v080r03 release.
Wow, nothing in 19 days. Anyway, I wanted to get Nick's logo back in
on the about screen.
Adds 80kb to both the binary and source archive, but eh. Gotta have
_some_ style. Nothing else new.
Probably won't be much news until the ST-0018 ROM is extracted.

Oh, and phoenix/Canvas sucks with GTK+, that'll be fixed eventually.
Hopefully prior to v081.