Update to bsnes v080r06 release.
Ran out of time, so this is incomplete, but ...

Windows will disable the compositor in fullscreen mode, and enable it
when switching back to windowed mode.
Should help with Vsync issues, but of course only in fullscreen mode.

I've also merged the four settings windows back into a panel with a
list view (since I have no tab control widget.)
The input settings window is a bit incomplete, need to break
assignment on window close, hide the capture buttons on first showing,
etc.
Will probably try and finish that up tonight.

[No archive available]