Update to bsnes v082r08 release.
Fixed up the PPU to be as close to cycle-perfect as possible. Fixed
RMW to write twice instead of read twice. Ryphecha added AOROM and
fixed up MMC1. Have CNROM too, but I need to rethink the mapper/board
distinction. Apparently the same logic IC is used in both AOROM and
CNROM, and it's just a matter of routing the pins to it. I need to
consider how crazy it'd be to emulate the logic IC and have boards
simply reroute pins to it. If it's too much work, we'll just treat
mappers as board + logic IC combinations. We'll see.