Update to bsnes v082r09 release.
Skip this build if you can, it's CPU+PPU timing improvements that
should not visibly affect anything.

Cheat codes and save states are going to work 100% the same between
every system. I'll leave the database for NES/GB up to mightymo, if he
is interested. They can even be in the same cheats.xml file, it
doesn't really matter. That's why XML is so nice, it can store any
form of cheat code for any system, no problem.