Update to bsnes v082r13 release.
I've updated the {System}::Interface classes to encapsulate all common
functionality, so they are C++ equivalents of libsnes now.
The idea being, use the interface class and you'll never have to reach
into core objects (unless you really want to.)
Not guaranteeing as stable an API as I do with libsnes for that,
though. C++ doesn't make for nice dynamic libraries, anyway.

Added back the state manager, and it now works for both SNES and Game
Boy. NES save states aren't in yet.
Anyway, this should give you a pretty good feel for what the combined
UI is going to be like: same as before, everything works the same.
Only difference is the dynamic system menu and cartridge menu with
more load options. The settings window will be mostly the same as
well, but will obviously have options that only apply to some systems.