Update to bsnes v082r23 release.
Ryphecha fixed the FF1 glitch, added two highpass filters to NES audio
output (still working on a lowpass), and fixed VRC6 audio issues.
I reduced the complexity of all eight supported mapping modes, and
standardized them; and added in an overscan filter (not in archive)
for chopping off all the NES edge garbage (8 pixels on the left and
right, 16 on the top and bottom.)
It's extreme, but anything less shows junk. I may make this part of
the menu option, just clip off more on NES mode than SNES mode.