Update to bsnes v082r25 release.
Ryphecha fixed Gun Nac, it was some sort of problem with blank sprite
address fetching messing with the MMC3
I've started on an XML parser for iNES-free loading, but it's pretty
barebones right now. Only NROM-256 loads, and you have to make it a
compile-time thing (so other games work for now.)
Updated nall with nullptr stuff.
nall/detect is now nall/intrinsic and has both #defines + static
constants that can be used to detect the platform (allows for run-time
platform checks where practical.)
ruby has a Makefile now, that makes using it in other projects a lot
easier

I need to rework my XML parser, it's not really all that suitable for
this, and NES is going to need probably 50x the XML parsing code of
the SNES.
I want to be able to quickly test if a required element exists or not,
and grab its data. I also want to design it around start/end offset
pairs, so that there's not so much malloc overhead.