Update to bsnes v086r12 release.
Attract demonstration game is fully playable.
---
Bingo. The F3 and F4 dumps are both valid.
Code:
	#include <stdlib.h>
	#include <string.h>

	//basically just add 32-bit words together. simple stuff.
	void checksum(const char *filename, unsigned size) {
	FILE *fp = fopen(filename, "rb");

	unsigned r0 = 0, r4, r9;
	while(ftell(fp) < size) {
	r4 = fgetc(fp);
	r9 = fgetc(fp);
	r4 |= r9 << 8;
	r9 = fgetc(fp);
	r4 |= r9 << 16;
	r9 = fgetc(fp);
	r4 |= r9 << 24;
	r0 += r4;
	}
	fclose(fp);
	printf("%s[0:%.5x] = %.8x\n", filename, size, r0);
	}

	int main() {
	checksum("0xf3.bin", 0x1f440);
	checksum("0xf4.bin", 0x05100);

	checksum("0xf3.bin", 0x20000);
	checksum("0xf4.bin", 0x08000);

	return 0;
	}
Code:
	0xf3.bin[0:1f440] = ecf28ee0
	0xf4.bin[0:05100] = bd8a23a5
	0xf3.bin[0:20000] = ecf28ee0
	0xf4.bin[0:08000] = a75e50f8
F3 has null data at the end of the file, thus its checksum is ecf28ee0 @ 1f440 and ecf28ee0 @ 20000.
F4 does not have null data at the end, thus its checksum is bd8a23a5 @ 5100 and a75e50f8 @ 8000.
So yes, random as it may appear ... both dumps are perfectly valid.
So I guess F4 is also ROM ... I'll rename it to dataROM for now.
For v087, I'm going to seed st0018.rom as the two ROMs appended together, like I do for the uPD chips and FEoEZ ROM. It'll be a 160KB file.
But until then, we'll leave it as st0018.rom and st0018d.rom, as we have it now.
And now with MUL/MLA supported, we're back to no invalid ops (that we know of) being executed ... so it's just some CPU bugs preventing the game from running fully.