Update to bsnes v087r03 release.
I wanted to keep this a secret, but unlike other recent additions,
this will easily take several weeks, maybe months, to show anything.
Assuming I can even pull it off. Nothing technically overwhelming
here, I'm more worried about the near-impossibility of debugging the
CPU.
So that would kill all WIPs until then, which would kind of suck.
Well, that's open source for you I guess.
Let's at least try and keep it a secret, please. No 'hinting', etc.
There's a good chance I'll give up on this, so we'll see.

If anyone is interested in co-developing this with me, I'd love the
help. A lot of stuff is unclear already.
So is the current CPU mode always stored in CPSR::M? Then what is
stored in the various SPSR::M values?

Current batshit insanity: in cpu.cpp, there is a line "apu.clock -=
clocks;".
If I uncomment that, the framerate drops from 720fps to 140fps.
Running -pg binaries is hanging on me, so I can't tell what the holy
fuck is going on there. But 4M subtractions a second should not cause
a 500% speed hit.
Would also appreciate any help in tracking down what is going on
there.