Update to bsnes v087r06 release.
I believe I've implemented every THUMB instruction now, although I'm
sure there are dozens of bugs in the implementation.

Here is a run-log: http://byuu.org/temp/armtrace.txt (instructions
_below_ current register values)
It seems that the last jump taken is ending up being off-by-two. It's
probably due to not masking/adjusting PC correctly at certain points.

I don't know if any other bugs are being hit prior to this or not.
I don't implement any I/O registers yet, and the BIOS seems to be
poking at a few of them along the way, so ... who knows.
I could also be reading the log wrong, but it looks to me like there's
some PSR setting the mode flag register to 0, which is supposed to be
an undefined behavior mode ... perhaps mrs has no effect on the m/t
bits, and it just affects the i/f bits?
It'd be nice if we could get a trace log like mine from a good
emulator running the BIOS to compare for issues.