Update to bsnes v087r07pre release.
Creates the bsnes/processor folder. This has a shared ARM core there
which both the GBA and ST018 inherit.
There are going to be separate decoders, and revision-specific checks,
to support the differences between v3+.
In the future, I also want to move the other processor cores here:
* GBZ80 (GB, GBC)
* 65816 (SNES CPU, SA-1)
* NEC uPD (7725, 96050, maybe 7720 just for fun)
* Hitachi HG51B169
* SuperFX
* SPC700
* 65(C?)02

Basically, the GBA/ST018 forces my hand to start coding a bit more
like a multi-system emulator.

Right now, the ST018 is broken. Hence the pre. Apparently the GBA core
being used now has some bugs. So this'll be a nice way to stress-test
the GBA core a bit before we make it to ARMwrestler.

[No archive available]