Update to bsnes v087r11 release.
Added all of the above fixes and changes.
Also new is pipeline_decode() to fetch data, and IME/IE/IF support,
and an ARM::processor.irqline flag that triggers IRQs at 0x18. Only
Vblank is hooked up, which is what SWI 4 was waiting on previously.
I'm sure my interrupt support is horribly broken and wrong. I was
never able to really figure out IE/IF on the Game Boy, so there's no
question this is even worse.
It's now going crazy and writing 0 to IE forever now after the Vblank
IRQ triggers.

Exophase, if you wouldn't mind taking a look and/or explaining IE/IF a
bit more, and how that all combines and updates CPU IRQ line state and
things like that, I'd appreciate it.
(Eg stuff like, if you clear IME does that flush IF to zero and reset
the CPU IRQ line instantly? Etc etc etc. The more detail, the better.)