Update to bsnes v087r23 release.
Changelog:
* fixed cascading timers and readouts (speed hit from 320fps to
240fps; would be 155fps with r20 timers) (fixes Spyro)
* OBJ mode 3 acts like OBJ mode 2 now (may not be correct, but nobody
has info on it)
* added background + object vertical+horizontal mosaic in all modes
(linear+affine+bitmap)
* object mosaic uses sprite (0,0) for start coordinates, not screen
(0,0) (again, nobody seems to have info on it)
* BIOS cannot be read by r(15)>=0x02000000; returns last BIOS read
instead (I can't believe games rely on this to work ... fixes SMA
Mario Bros.)

Mosaic is what concerns me the most, I've no idea if I'm doing it
correctly. But anything is probably better than nothing, so there's
that.
I don't really notice the effect in Metroid Fusion. So either it's
broken, or it's really subtle.