Update to bsnes v087r28 release.
Be sure to run make install, and move required images to their
appropriate system profile folders.
I still have no warnings in place if those images aren't present.

Changelog:
* OBJ mosaic should hopefully be emulated correctly now (thanks to
krom and Cydrak for testing the hardware behavior)
* emulated dummy serial registers, fixes Sonic Advance (you may still
need to specify 512KB FlashROM with an appropriate ID, I used
Panaonic's)
* GBA core exits scheduler (PPU thread) and calls
interface->videoRefresh() from main thread (not required, just nice)
* SRAM, FRAM, EEPROM and FlashROM initialized to 0xFF if it does not
exist (probably not needed, but FlashROM likes to reset to 0xFF
anyway)
* GBA manifest.xml for file-mode will now use "gamename.xml" instead
of "gamename.gba.xml"
* started renaming "NES" to "Famicom" and "SNES" to "Super Famicom" in
the GUI (may or may not change source code in the long-term)
* removed target-libsnes/
* added profile/

Profiles are the major new feature. So far we have:
    Famicom.sys/{nothing (yet?)}
    Super Famicom.sys/{ipl.rom}
    Game Boy.sys/{boot.rom}
    Game Boy Color.sys/{boot.rom}
    Game Boy Advance.sys/{bios.rom[not included]}

    Super Game Boy.sfc/{boot.rom,program.rom[not included]}
    BS-X Satellaview.sfc/{program.rom,bsx.ram,bsx.pram}
    Sufami Turbo.sfc/{program.rom}


The SGB, BSX and ST cartridges ask you to load GB, BS or ST cartridges
directly now. No slot loader for them.
So the obvious downsides: you can't quickly pick between different SGB
BIOSes, but why would you want to? Just use SGB2/JP.
It's still possible, so I'll sacrifice a little complexity for a rare
case to make it a lot easier for the more common case.
ST cartridges currently won't let you load the secondary slot.
BS-X Town cart is the only useful game to load with nothing in the
slot, but only barely, since games are all seeded on flash and not on
PSRAM images. We can revisit a way to boot the BIOS directly if and
when we get the satellite uplink emulated and data can be downloaded
onto the PSRAM :P
BS-X slotted cartridges still require the secondary slot.

My plan for BS-X slotted cartridges is to require a manifest.xml to
specify that it has the BS-X slot present.
Otherwise, we have to load the ROM into the SNES cartridge class, and
parse its header before we can find out if it has one. Screw that.
If it's in the XML, I can tell before loading the ROM if I need to
present you with an optional slot loading dialog.
I will probably do something similar for Sufami Turbo. Not all games
even work with a secondary slot, so why ask you to load a second slot
for them? Let the XML request a second slot. A complete Sufami Turbo
ROM set will be trivial anyway.
Not sure how I want to do the sub dialog yet. We want basic file
loading, but we don't want it to look like the dialog 'didn't do
anything' if it pops back open immediately again. Maybe change the
background color of the dialog to a darker gray? Tacky, but it'd give
you the visual cue without the need for some subtle text changes.