Update to bsnes v088r02 release.
Really no reason to download this one unless you want to stay on the
absolute bleeding edge of phoenix development.

Basically, the current implementation of nall/array is deprecated now.
The old method was for non-reference types, it acted like a vector for
POD types (raw memory allocation instead of placement new
construction.)
And for reference types, it acted like an unordered set. Yeah, not
good.

As of right now, nall/array is no longer used. The vector type usage
was replaced with actual vectors.
I've created nall/set, which now contains the specialization for
reference types.
nall/set basically acts much like std::unordered_set. No auto-sort,
only one of each type is allowed, automatic growth.
This will be the same both for reference and non-reference types ...
however, the non-reference type wasn't implemented just yet.
Future plans for nall/array are for it to be a statically allocated
block of memory, ala array<type, size>, which is meant for RAII memory
usage.
Have to work on the specifics, eg the size as a template parameter may
be problematic. I'd like to return allocated chunks of memory (eg
file::read) in this container so that I don't have to manually free
the data anymore.

I also removed nall/moduloarray, and moved that into the SNES DSP
class, since that's the only thing that uses it.