Update to bsnes v088r12 release.
Changelog:
* all hotkeys from target-ui now exist in target-ethos
* controller port menus now show up when you load a system (hidden if
there are no options to choose from)
* tools menu auto-hides with no game open ... not much point to it
then
* since we aren't using RawInput's multi-KB/MS support anyway, input
and hotkey mappings remove KB0:: and turn MS0:: into Mouse::, makes it
a lot easier to read
* added mute audio, sync video, sync audio, mask overscan
* added video settings: saturation, gamma, luminance, overscan
horizontal, overscan vertical
* added audio settings: frequency, latency, resampler, volume
* added input settings: when focus is lost [ ] pause emulator [ ]
allow input
* pausing and autopausing works
* status messages hooked up (show a message in status bar for a few
seconds, then revert to normal status text)
* sub systems (SGB, BSX, ST) sorted below primary systems list
* added geometry settings cache
* Emulator::Interface cleanups and simplifications
* save states go into (cart foldername.extension/bsnes/state-#.bsa)
now. Idea is to put emulator-specific data in their own subfolders

Caveats / Missing:
~~* SGB input does not work~~
~~* Sufami Turbo second slot doesn't work yet~~
* BS-X BIOS won't show the data pack
~~* NEC/Hitachi/ARM DSP firmware not loading~~
~~* need to save RAM files to their own slot folders~~
* need XML mapping information window
* need cheat editor and cheat database
~~* need state manager~~
* need video shaders
* need driver selection
* need NSS DIP switch settings
* need to hide controllers that have no inputs from the input mapping
list

So for video settings, I used to have contrast/brightness/gamma.
Contrast was just a multiplier on intensity of each channel, and
brightness was an addition or subtraction against each channel. They
kind of overlapped and weren't that effective. The new setup has
saturation, gamma and luminance.

Saturation of 100% is normal. If you lower it, color information goes
away. 0% = grayscale. If you raise it, color intensity increases (and
clamps.)
This is wonderful for GBA games, since they are oversaturated to
fucking death. Of course we'll want to normalize that inside the core,
so the same sat. value works on all systems, but for now it's nice. If
you raise saturation above 100%, it basically acts like contrast used
to. It's just that lowering it fades to grayscale rather than black.
Adding doesn't really work well for brightness, it throws off the
relative distance between channels and looks like shit.
So now we have luminance, which takes over the old contrast <100%
role, and just fades the pixels toward black. Obviously, luminance >
100% would be the same as saturation > 100%, so that isn't allowed, it
caps at 100% now.
Gamma's the same old function. Gamma curve on the lower-half of the
color range.
Effects are applied in the order they appear in the GUI: color ->
saturate -> gammify -> luminate -> output.